+++ src/extension.ts
@@ -7,8 +7,15 @@ import { Runner } from './runner';
 export function activate(context: vscode.ExtensionContext) {
     // Use the console to output diagnostic information (console.log) and errors (console.error).
     // This line of code will only be executed once when your extension is activated.
-    let runner = new Runner();
-    let controller = new Controller(runner);
+    const runner = new Runner();
+    // tslint:disable-next-line:no-use-before-declare
+    const controller = new Controller(runner);
+
+    // Disable matchBrackets
+    const disableNative = vscode.workspace.getConfiguration().get<boolean>('subtleBrackets.disableNative');
+    if (disableNative) {
+        vscode.workspace.getConfiguration().update('editor.matchBrackets', false, true);
+    }
 
     // Add to a list of disposables which are disposed when this extension is deactivated.
     context.subscriptions.push(controller, runner);
@@ -18,12 +25,12 @@ class Controller {
     private _runner: Runner;
     private _disposable: vscode.Disposable;
 
-    constructor(runner: Runner) {
+    constructor (runner: Runner) {
         this._runner = runner;
         this._runner.run();
 
         // Subscribe to selection change and editor activation events
-        let subscriptions: vscode.Disposable[] = [];
+        const subscriptions: vscode.Disposable[] = [];
         vscode.window.onDidChangeTextEditorSelection(this._onEvent, this, subscriptions);
         vscode.window.onDidChangeActiveTextEditor(this._onEvent, this, subscriptions);
 
@@ -31,7 +38,7 @@ class Controller {
         this._disposable = vscode.Disposable.from(...subscriptions);
     }
 
-    dispose() {
+    public dispose() {
         this._disposable.dispose();
     }