+++ src/runner.ts
@@ -1,48 +1,64 @@
 'use strict';
 import * as vscode from 'vscode'; // VS Code extensibility API
+import * as prism from 'prismjs';
+import * as prismLanguages from 'prism-languages';
 
-const Prism = require('prismjs'),
-    prismLanguages = require('prism-languages');
+interface IKeepParsingObj {
+    open: number;
+    aBracket: string;
+    bBracket: string;
+    line: number;
+    lineText: string;
+    direction: number;
+    doc: vscode.TextDocument;
+    lineCount: number;
+    knownLanguage: string;
+    typesFound: string[];
+    parsedDocBrackets: object;
+}
 
 export class Runner {
 
-    private bracketsDict: {'all': string[], 'open': string[], 'close': string[], 'pairs': {}};
+    private bracketsDict: { all: string[], open: string[], close: string[], pairs: {} };
     private decoration: vscode.TextEditorDecorationType;
     private past: boolean;
     private languages: string[];
     private regexp: RegExp;
-    private parse: Boolean;
+    private parse: boolean;
 
-    constructor() {
+    constructor () {
 
-        const escape = (s) => s.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&'), // FIXME First parenthesis doesn't work with subtle brackets
-            config = vscode.workspace.getConfiguration('subtleBrackets');
+        const escape = (s) => s.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&');
+        const config = vscode.workspace.getConfiguration('subtleBrackets');
 
-        let style = config.style;
+        const style = config.style;
         if (!config.style.hasOwnProperty('borderColor')) {
             style.borderColor = '#D4D4D4';
-            style.light = { 'borderColor': '#333333' };
+            style.light = { borderColor: '#333333' };
         }
 
         this.decoration = vscode.window.createTextEditorDecorationType(style);
         this.past = false;
-        this.bracketsDict = {'all': [], 'open': [], 'close': [], 'pairs': {}};
+        this.bracketsDict = {all: [], open: [], close: [], pairs: {}};
         this.languages = Object.keys(prismLanguages);
         this.parse = config.parse;
 
         config.bracketPairs.forEach(x => {
             if (x.length === 2) { // Safety Check
-                let [open, close] = x.split('');
-                this.bracketsDict['all'].push(open, close);
-                this.bracketsDict['open'].push(open);
-                this.bracketsDict['close'].push(close);
-                this.bracketsDict['pairs'][open] = close;
-                this.bracketsDict['pairs'][close] = open;
+                const [open, close] = x.split('');
+                this.bracketsDict.all.push(open, close);
+                this.bracketsDict.open.push(open);
+                this.bracketsDict.close.push(close);
+                this.bracketsDict.pairs[open] = close;
+                this.bracketsDict.pairs[close] = open;
             }
         });
 
-        this.regexp = new RegExp('[' + escape(this.bracketsDict['all'].join('')) + ']', 'g');
+        this.regexp = new RegExp('[' + escape(this.bracketsDict.all.join('')) + ']', 'g');
     }
+
+    // PUBLIC
+    public dispose() {  }
 
     public run() {
 
@@ -56,36 +72,36 @@ export class Runner {
             this.past = false;
         }
 
-        const doc = editor.document,
-            selection = vscode.window.activeTextEditor.selection,
-            range = new vscode.Range(selection.start, selection.end);
+        const doc = editor.document;
+        const selection = vscode.window.activeTextEditor.selection;
+        const range = new vscode.Range(selection.start, selection.end);
 
         if (!range.isEmpty) return;
-        // There's no selection, only the cursor on a position
+        // Fallthrough: Range empty, there's no selection, only the cursor on a position
 
-        const line = selection.start.line,
-            startChar = selection.start.character,
-            lineText = doc.lineAt(line).text,
-            prevChar = Math.max(startChar-1, 0),
-            postChar = Math.min(startChar+1, lineText.length),
-            // Left and right characters to the cursor
-            left = lineText.slice(prevChar, startChar),
-            right = lineText.slice(startChar, postChar),
-            // Booleans, checking whether they are brackets
-            isRight = right && this.bracketsDict.all.indexOf(right) !== -1,
-            isLeft = left && this.bracketsDict.all.indexOf(left) !== -1;
+        const line = selection.start.line;
+        const startChar = selection.start.character;
+        const lineText = doc.lineAt(line).text;
+        const prevChar = Math.max(startChar-1, 0);
+        const postChar = Math.min(startChar+1, lineText.length);
+        // Left and right characters to the cursor
+        const left = lineText.slice(prevChar, startChar);
+        const right = lineText.slice(startChar, postChar);
+        // Booleans, checking whether they are brackets
+        const isRight = right && this.bracketsDict.all.indexOf(right) !== -1;
+        const isLeft = left && this.bracketsDict.all.indexOf(left) !== -1;
 
         if (!(isRight || isLeft)) return; // If they're not brackets, return
 
-        const [knownLanguage, parsedDocBrackets, typesFound] = this.parseDocBrackets(doc),
-            lineTillLeft = lineText.slice(0, prevChar),
-            lineTillRight =  lineText.slice(0, startChar);
+        const {knownLanguage, parsedDocBrackets, typesFound} = this.parseDocBrackets(doc);
+        const lineTillLeft = lineText.slice(0, prevChar);
+        const lineTillRight =  lineText.slice(0, startChar);
 
         // Let's see if any of `left` and `right` are brackets
         // Prioritizing `right`
-        let aBracket: string,
-            startPosChar: number,
-            endPosChar: number;
+        let aBracket: string;
+        let startPosChar: number;
+        let endPosChar: number;
         if (isRight && this.isPunctuationType(knownLanguage, typesFound,
                                         parsedDocBrackets, line, right, lineTillRight)) {
             aBracket = right;
@@ -102,14 +118,14 @@ export class Runner {
         }
 
         // Range of the selected bracket
-        let aBracketRange = new vscode.Range(
-                new vscode.Position(line, startPosChar),
-                new vscode.Position(line, endPosChar)
-            );
+        const aBracketRange = new vscode.Range(
+            new vscode.Position(line, startPosChar),
+            new vscode.Position(line, endPosChar)
+        );
 
         // Let's see if the bracket is an opening or closing bracket
-        let lineRest: string,
-            parseDirection: number;
+        let lineRest: string;
+        let parseDirection: number;
         if (this.bracketsDict.open.indexOf(aBracket) !== -1) {
             // Opening bracket. Parse forwards
             lineRest = lineText.slice(endPosChar);
@@ -121,10 +137,23 @@ export class Runner {
         }
 
         // Complement of `aBracket`
-        let bBracket = this.bracketsDict.pairs[aBracket],
-            [bLine, bChar] = this.parseUntilComplement(1, aBracket, bBracket, line,
-                                lineRest, parseDirection, doc, doc.lineCount,
-                                knownLanguage, typesFound, parsedDocBrackets);
+        const bBracket = this.bracketsDict.pairs[aBracket];
+        const kpObj = {
+            open: 1,
+            aBracket: aBracket,
+            bBracket: bBracket,
+            line: line,
+            lineText: lineRest,
+            direction: parseDirection,
+            doc: doc,
+            lineCount: doc.lineCount,
+            knownLanguage: knownLanguage,
+            typesFound: typesFound,
+            parsedDocBrackets: parsedDocBrackets
+        };
+
+        // tslint:disable-next-line:prefer-const
+        let [bLine, bChar] = this.parseUntilComplement(kpObj);
 
         // Verify the parsing didn't fail
         if (bLine < 0) return;
@@ -133,49 +162,61 @@ export class Runner {
         if (line === bLine && parseDirection > 0) bChar += endPosChar;
 
         // Now we have `aBracketRange` and `bBracketRange`
-        let bBracketRange = new vscode.Range(
-                new vscode.Position(bLine, bChar),
-                new vscode.Position(bLine, bChar+bBracket.length)
-            );
+        const bBracketRange = new vscode.Range(
+            new vscode.Position(bLine, bChar),
+            new vscode.Position(bLine, bChar+bBracket.length)
+        );
         this.past = true;
         editor.setDecorations(this.decoration, [bBracketRange, aBracketRange]);
     }
 
-    private parseUntilComplement(open: number, aBracket: string, bBracket: string,
-                    line: number, lineText: string, direction: number,
-                    doc: vscode.TextDocument, lineCount: number, knownLanguage: string,
-                    typesFound: string[], parsedDocBrackets: Object): number[] {
+    // PRIVATE
+
+    private parseUntilComplement(kpObj: IKeepParsingObj, first: boolean = true): number[] {
         let lastChar: number;
 
-        for (let ii = 0; ii < lineText.length; ii++) {
-            if (open === 0) break;
+        for (let ii = 0; ii < kpObj.lineText.length; ii++) {
+            if (kpObj.open === 0) break;
             // Parse line backwards if looking for opening bracket
-            if (direction < 0) lastChar = lineText.length - ii - 1;
-            else lastChar = ii;
+            lastChar = (kpObj.direction < 0) ? (kpObj.lineText.length - ii - 1) : ii;
             // Add or remove brackets to open/close
-            let char = lineText[lastChar],
-                lineTill = lineText.slice(0, lastChar),
-                isA = char === aBracket,
-                isB = char === bBracket;
+            const char = kpObj.lineText[lastChar];
+            let lineTill: string;
+            if (first && kpObj.direction > 0) {
+                lineTill = kpObj.doc.lineAt(kpObj.line).text;
+                lineTill = lineTill
+                    .slice(0, lastChar + (lineTill.length - kpObj.lineText.length));
+            } else {
+                lineTill = kpObj.lineText.slice(0, lastChar);
+            }
+            const [isA, isB] = [char === kpObj.aBracket, char === kpObj.bBracket];
 
-            if ((isA || isB) && this.isPunctuationType(knownLanguage, typesFound,
-                                            parsedDocBrackets, line, char, lineTill)) {
-                if (isA) open++;
-                else if (isB) open--;
+            if ((isA || isB) && this.isPunctuationType(kpObj.knownLanguage,
+                kpObj.typesFound, kpObj.parsedDocBrackets, kpObj.line, char, lineTill)) {
+                if (isA) kpObj.open++;
+                else if (isB) kpObj.open--;
             }
         }
-        if (open === 0) return [line, lastChar]; // Found Complement
+        if (kpObj.open === 0) {
+            // Found Complement
+            return [kpObj.line, lastChar];
+        }
+
         // Not found
-        line = line+direction;
-        if (line >= lineCount || line < 0) return [-1, -1]; // Out of bounds
+        kpObj.line = kpObj.line + kpObj.direction;
+        if (kpObj.line >= kpObj.lineCount || kpObj.line < 0) {
+            // Out of bounds
+            return [-1, -1];
+        }
+
         // Keep looking on next line
-        lineText = doc.lineAt(line).text;
-        return this.parseUntilComplement(open, aBracket, bBracket, line, lineText,
-                 direction, doc, lineCount, knownLanguage, typesFound, parsedDocBrackets);
+        kpObj.lineText = kpObj.doc.lineAt(kpObj.line).text;
+        return this.parseUntilComplement(kpObj, false);
     }
 
     private isPunctuationType(knownLanguage: string, typesFound: string[],
-            parsedDocBrackets: Object, line: number, bracket: string, lineTill: string) {
+        parsedDocBrackets: object, line: number, bracket: string, lineTill: string) {
+
         if ((!knownLanguage) || ((!typesFound.indexOf(bracket))
                         && (!typesFound.indexOf(this.bracketsDict.pairs[bracket])))) {
             // Decorate all if language is not known or no bracket at all of that pair
@@ -183,32 +224,37 @@ export class Runner {
             return true;
         }
 
-        const nBracketInLine = (text, bracket) => text.split(bracket).length - 1;
+        const nBracketInLine = (text, brkt) => text.split(brkt).length - 1;
         if (parsedDocBrackets.hasOwnProperty(line)
                             && parsedDocBrackets[line].hasOwnProperty(bracket)) {
             const char = nBracketInLine(lineTill, bracket);
             return parsedDocBrackets[line][bracket][char] === 'punctuation';
-        } else return true; // Didn't find it in the parsed obj; swallow error and return true anyways
+        }
+
+        // Didn't find it in the parsed obj; swallow error and return true anyways
+        return true;
     }
 
-    private parseDocBrackets(doc): any[] {
-        const emptyAns = ['', {}, []],
-            language = doc.languageId;
-        if ((!this.parse) || (!this.languages.indexOf(language))) return emptyAns;
-
+    private parseDocBrackets(doc): {knownLanguage: string, parsedDocBrackets: object,
+                                                                typesFound: string[]} {
+        // Helper Function
         const getContent = (lines, typesFound, tokenized, currentLine, currentLineAt) => {
             tokenized.forEach(token => {
                 if (typeof token === 'string') {
-                    let splits = token.split('\n');
+                    const splits = token.split('\n');
                     currentLine = currentLine + splits.length-1;
                     currentLineAt = splits.slice(-1)[0].length;
                 }
                 else if (typeof token.content === 'string') {
-                    const currentContent = token.content,
-                        matches = currentContent.match(this.regexp);
+                    const currentContent = token.content;
+                    const matches = currentContent.match(this.regexp);
                     if (matches) matches.forEach(bracket => {
-                        if (!lines.hasOwnProperty(currentLine)) lines[currentLine] = {};
-                        if (!lines[currentLine].hasOwnProperty(bracket)) lines[currentLine][bracket] = [];
+                        if (!lines.hasOwnProperty(currentLine)) {
+                            lines[currentLine] = {};
+                        }
+                        if (!lines[currentLine].hasOwnProperty(bracket)) {
+                            lines[currentLine][bracket] = [];
+                        }
                         lines[currentLine][bracket].push(token.type);
                         typesFound[bracket] = true;
                     });
@@ -222,24 +268,28 @@ export class Runner {
                 }
             });
             return [lines, typesFound, currentLine, currentLineAt, currentLineAt];
-        }
+        };
+
+        const emptyAns = { knownLanguage: '', parsedDocBrackets: {}, typesFound: [] };
+        const language = doc.languageId;
+        if ((!this.parse) || (!this.languages.indexOf(language))) return emptyAns;
 
         // Known language, parse with Prism
         const fullText = doc.getText();
         let tokenized;
         try {
-            tokenized = Prism.tokenize(fullText, Prism.languages[language]);
+            tokenized = prism.tokenize(fullText, prism.languages[language]);
         } catch(err) {
             return emptyAns;
         }
         if (!tokenized) return emptyAns;
 
-        const ansGetContent = getContent({}, {}, tokenized, 0, 0),
-            lines = ansGetContent[0],
-            typesFound = Object.keys(ansGetContent[1]);
-        return [language, lines, typesFound];
+        const ansGetContent = getContent({}, {}, tokenized, 0, 0);
+        return {
+            knownLanguage: language,
+            parsedDocBrackets: ansGetContent[0],
+            typesFound: Object.keys(ansGetContent[1])
+        };
     }
 
-    dispose() {
-    }
 }