+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to *Subtle Brackets* will be documented in this file.
 
+## [1.1.0] 2017-06-12
+- [Prism](http://prismjs.com/) to stable: fix errors
+- Set `matchBrackets` to `false` automatically
+
 ## [1.0.8] 2017-06-09
 - README: Settings to table
 - Fix for VSCode v1.13.0: ensure automatic border color for dark and light themes